new up

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: moer-ret <moer-ret@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/29 02:46:19 by moer-ret          #+#    #+#             */
-/*   Updated: 2023/11/30 22:16:00 by moer-ret         ###   ########.fr       */
+/*   Updated: 2023/12/02 03:14:30 by moer-ret         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,13 +15,14 @@
 
 # include <unistd.h>
 # include <stdlib.h>
+# include <limits.h>
 # include <stdarg.h>
 
 int	ft_printf(const char *format, ...);
 int	ft_putchar(char c);
 int	ft_putstr(char *str);
 int	ft_pointer(unsigned long nb);
-int	ft_putnbr(long int n);
+int	ft_putnbr(int n);
 int	ft_putnbr_unsigned(unsigned int n);
 int	ft_puthex(unsigned int nb, const char format);
